+++ api/src/main/java/com/percolate/sdk/api/request/campaignsection/CampaignSectionsRequest.java
@@ -1,4 +1,4 @@
-package com.percolate.sdk.rxjava.request.campaignsection;
+package com.percolate.sdk.api.request.campaignsection;
 
 import com.percolate.sdk.api.PercolateApi;
 import com.percolate.sdk.api.request.campaign.CampaignSectionsListParams;
@@ -10,27 +10,27 @@ import com.percolate.sdk.dto.SingleCampaignSection;
 import org.jetbrains.annotations.NotNull;
 
 import okhttp3.ResponseBody;
-import rx.Observable;
+import retrofit2.Call;
 
 /**
  * Campaign Sections request proxy.
  */
 @SuppressWarnings("unused")
-public class CampaignSectionRequestRx {
+public class CampaignSectionsRequest {
 
-    private CampaignSectionServiceRx service;
+    private CampaignSectionService service;
 
-    public CampaignSectionRequestRx(@NotNull PercolateApi context) {
-        this.service = new RetrofitApiFactory(context).getService(CampaignSectionServiceRx.class);
+    public CampaignSectionsRequest(@NotNull PercolateApi context) {
+        this.service = new RetrofitApiFactory(context).getService(CampaignSectionService.class);
     }
 
     /**
      * Get a campaign section.
      *
      * @param campaignSectionId Campaign section id.
-     * @return {@link Observable} object.
+     * @return {@link Call} object.
      */
-    public Observable<SingleCampaignSection> get(@NotNull final String campaignSectionId) {
+    public Call<SingleCampaignSection> get(@NotNull final String campaignSectionId) {
         return service.get(campaignSectionId);
     }
 
@@ -38,9 +38,9 @@ public class CampaignSectionRequestRx {
      * Get campaign sections.
      *
      * @param params API params.
-     * @return {@link Observable} object.
+     * @return {@link Call} object.
      */
-    public Observable<CampaignSections> list(@NotNull final CampaignSectionsListParams params) {
+    public Call<CampaignSections> list(@NotNull final CampaignSectionsListParams params) {
         return service.list(params.getParams());
     }
 
@@ -48,9 +48,9 @@ public class CampaignSectionRequestRx {
      * Create campaign section.
      *
      * @param campaignSectionData {@link CampaignSectionData} object.
-     * @return {@link Observable} object.
+     * @return {@link SingleCampaignSection} object.
      */
-    public Observable<SingleCampaignSection> create(@NotNull final CampaignSectionData campaignSectionData) {
+    public Call<SingleCampaignSection> create(@NotNull final CampaignSectionData campaignSectionData) {
         return service.create(campaignSectionData);
     }
 
@@ -58,9 +58,9 @@ public class CampaignSectionRequestRx {
      * Update campaign section.
      *
      * @param campaignSectionData {@link CampaignSectionData} object.
-     * @return {@link Observable} object.
+     * @return {@link SingleCampaignSection} object.
      */
-    public Observable<SingleCampaignSection> update(@NotNull final CampaignSectionData campaignSectionData) {
+    public Call<SingleCampaignSection> update(@NotNull final CampaignSectionData campaignSectionData) {
         return service.update(campaignSectionData.getId(), campaignSectionData);
     }
 
@@ -68,9 +68,9 @@ public class CampaignSectionRequestRx {
      * Delete campaign section.
      *
      * @param campaignSectionId Campaign section id.
-     * @return {@link Observable} object.
+     * @return {@link ResponseBody} object.
      */
-    public Observable<ResponseBody> delete(@NotNull final String campaignSectionId) {
+    public Call<ResponseBody> delete(@NotNull final String campaignSectionId) {
         return service.delete(campaignSectionId);
     }
 }
